moved activitiesUpdater test in MainMapFragmentTest to test dependency
