+++ app/src/androidTest/java/com/sdp/movemeet/utility/ActivitiesUpdaterTest.java
@@ -11,6 +11,7 @@ import com.google.firebase.firestore.FirebaseFirestore;
 import com.sdp.movemeet.Activity.Activity;
 import com.sdp.movemeet.Backend.BackendActivityManager;
 import com.sdp.movemeet.Sport;
+import com.sdp.movemeet.map.MainMapFragment;
 
 import org.junit.Before;
 import org.junit.Rule;
@@ -43,50 +44,6 @@ public class ActivitiesUpdaterTest {
     @Test
     public void instanceIsNeverNull() {
         assertNotNull(updater);
-    }
-
-    @Test
-    public void activitiesUpdatesOnAdd() {
-
-        Activity act = new Activity("activity",
-                "me",
-                "title",
-                10,
-                new ArrayList<String>(),
-                0,
-                0,
-                "desc",
-                new Date(),
-                10,
-                Sport.Running,
-                "address",
-                new Date());
-
-        bam.uploadActivity(act,
-                new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void aVoid) {
-                        List<Activity> activities = updater.getActivities();
-                        Activity act_in_collection = activities.get(activities.size()-1);
-                        assertEquals(act.getActivityId(), act_in_collection.getActivityId());
-                        assertEquals(act.getAddress(), act_in_collection.getAddress());
-                        assertEquals(act.getDate(), act_in_collection.getDate());
-                        assertEquals(act.getDescription(), act_in_collection.getDescription());
-                        //assertEquals(act.getLatitude(), act_in_collection.getLatitude());
-                        //assertEquals(act.getLongitude(), act_in_collection.getLongitude());
-                        assertEquals(act.getNumberParticipant(), act_in_collection.getNumberParticipant());
-                        assertEquals(act.getParticipantId(), act_in_collection.getParticipantId());
-                        assertEquals(act.getTitle(), act_in_collection.getTitle());
-                        assertEquals(act.getOrganizerId(), act_in_collection.getOrganizerId());
-                        assertEquals(act.getSport(), act_in_collection.getSport());
-                    }
-                },
-                new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                    }
-                });
-
     }
 
 }
